refactor: simplify images in prep for finishing menu
