+++ routes/menu.tsx
@@ -1,9 +1,6 @@
 import { type PageResponse, page } from "fresh";
 import type { VNode } from "preact";
-import {
-  MenuImage,
-  type MenuImageProps,
-} from "../islands/images/MenuImage.tsx";
+import { Image, type MenuImageProps } from "../islands/Image.tsx";
 import { define } from "../utils.ts";
 
 export const handler = define.handlers({
@@ -33,7 +30,7 @@ export default define.page<typeof handler>(() => {
           description="Delicious gluten free pancakes filled with poppy seeds and oats. Topped with a berry chia jam."
           side="left"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -42,7 +39,7 @@ export default define.page<typeof handler>(() => {
           description="Hearty waffles made with almonds, oats, applesauce, and sesame seeds. Topped with mixed berries and syrup."
           side="right"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -51,7 +48,7 @@ export default define.page<typeof handler>(() => {
           description="Thin rice based crepes filled with wild herbs and homemade cashew parmesan cheese."
           side="left"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -60,7 +57,7 @@ export default define.page<typeof handler>(() => {
           description="There's never a bad time for burgers! Millet and kidney bean based patties topped with tofu, served on a whole grain bun."
           side="right"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -69,7 +66,7 @@ export default define.page<typeof handler>(() => {
           description="Who needs eggs for omelets when you have tofu? Our tofu omelets are stuffed with slowly roasted tomatoes for a scrumptious, healthy meal."
           side="left"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -78,7 +75,7 @@ export default define.page<typeof handler>(() => {
           description="Sleek and delicious tofu garnished with the finest chives and served with rye bread."
           side="right"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
       </div>
 
@@ -93,7 +90,7 @@ export default define.page<typeof handler>(() => {
           description="Satisfy your morning sweet tooth with sugary carrot cupcakes topped with lemon coconut frosting."
           side="left"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -102,7 +99,7 @@ export default define.page<typeof handler>(() => {
           description="Raised glazed donuts with a chocolate glaze topped with hazelnuts and pistachios."
           side="right"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -111,7 +108,7 @@ export default define.page<typeof handler>(() => {
           description="Lemon muffins filled with chia seeds and coconut. A great way to start your day!"
           side="left"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -120,7 +117,7 @@ export default define.page<typeof handler>(() => {
           description="An almond and coconut base topped with a lemon, cashew, and strawberry filling."
           side="right"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
       </div>
 
@@ -133,7 +130,7 @@ export default define.page<typeof handler>(() => {
           description="Kiwis, bananas, grapefruit, strawberries, pineapple, blueberries, and raspberries—oh my! Served with a delightful coconut pudding."
           side="left"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -142,7 +139,7 @@ export default define.page<typeof handler>(() => {
           description="If only goldilocks tried this porridge first—its blend of millet, quinoa, cinnamon, and fresh fruit would have been just right!"
           side="right"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -151,7 +148,7 @@ export default define.page<typeof handler>(() => {
           description="Banana, mango, kiwi, pineapple, and spinach, all blended together and topped with raspberries and chia seeds."
           side="left"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -160,7 +157,7 @@ export default define.page<typeof handler>(() => {
           description="Our mix of banana, raspberries, blackberries, cranberries, spinach, and oats makes for a berry delicious breakfast."
           side="right"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -169,7 +166,7 @@ export default define.page<typeof handler>(() => {
           description="Oats, peas, soy sauce, and white miso cooked together to make one tasty meal."
           side="left"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
       </div>
 
@@ -184,7 +181,7 @@ export default define.page<typeof handler>(() => {
           description="Whole wheat tortillas stuffed with tomatoes, avocados, oats, and chickpeas topped with mushrooms."
           side="left"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -193,7 +190,7 @@ export default define.page<typeof handler>(() => {
           description="Tempeh bacon, romaine lettuce, and tomato between a sturdy hash brown bun."
           side="right"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -202,7 +199,7 @@ export default define.page<typeof handler>(() => {
           description="Experience the life of English royalty with a full English breakfast of baked beans, mushrooms, vegan sausage, tempeh bacon, baby spinach, grilled tomatoes, and chia toast."
           side="left"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
       </div>
 
@@ -217,7 +214,7 @@ export default define.page<typeof handler>(() => {
           description="Warm, fragrant ginger tea with a hint of lemon."
           side="left"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -226,7 +223,7 @@ export default define.page<typeof handler>(() => {
           description="Smooth and rich cold brew coffee, steeped for 24 hours."
           side="right"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
 
         <MenuItem
@@ -235,7 +232,7 @@ export default define.page<typeof handler>(() => {
           description="Creamy oat milk blended with premium ceremonial-grade matcha."
           side="left"
         >
-          <MenuImage image="what-we-serve_r9qi3z" description="Yum!" />
+          <Image id="what-we-serve_r9qi3z" description="Yum!" />
         </MenuItem>
       </div>
     </main>
@@ -249,7 +246,7 @@ interface MenuItemProps {
 
   side: "left" | "right";
 
-  /** An instance of {@linkcode MenuImage}. */
+  /** An instance of {@linkcode Image}. */
   children: VNode<MenuImageProps>;
 }
